Build FAMs with sentinel byte and casting
I originally wanted to use extensions to make FAMs. There were three choices: C 99-style, array of length 0, or array of length 1 where we write past the end. Not all of these are supported under all compilers and compile modes, and supporting all of them created a lot of complexity, especially when it came to initialization.
Instead, put a byte aligned like a T at the end of a struct. This gives us the correct address the FAM starts at. As long as memory was allocated properly, an implicit array of uninitialized Ts lives there.